updated schema to use big int to store tweet id

+++ schema/calvin.sql
@@ -3,7 +3,7 @@
 -- http://www.phpmyadmin.net
 --
 -- Host: localhost
--- Generation Time: Jan 11, 2016 at 12:57 PM
+-- Generation Time: Jan 12, 2016 at 03:07 PM
 -- Server version: 5.5.46-0ubuntu0.14.04.2
 -- PHP Version: 5.5.9-1ubuntu4.14
 
@@ -29,7 +29,7 @@ SET time_zone = "+00:00";
 CREATE TABLE IF NOT EXISTS `media` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `url` varchar(2000) NOT NULL,
-  `tweet_id` int(64) NOT NULL,
+  `tweet_id` bigint(20) NOT NULL,
   `path` text NOT NULL,
   PRIMARY KEY (`id`),
   KEY `tweet_id` (`tweet_id`)
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS `media` (
 --
 
 CREATE TABLE IF NOT EXISTS `tweet` (
-  `id` int(64) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `text` varchar(140) NOT NULL,
   `created_at` datetime NOT NULL,
   `inserted_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
